Make it possible to choose what to build
By specifying -DBUILD_<COMPONENT>=False on the cmake command line, individual
components can be switched off.

Available components are:

  EMULATOR
  SERVER
  PIPELINE

Disabling a components means that its dependencies are no longer checked for.
Since PIPELINE is the only component which depends on Matlab, it is this
possible to build EMULATOR and SERVER on a machine with no Matlab installed.